+++ apps/gateway/src/routes/api/v3/projects/versions/documents/index.ts
@@ -9,6 +9,10 @@ import {
   createOrUpdateDocumentRouteConfig,
 } from '$/routes/api/v3/projects/versions/documents/createOrUpdate'
 import {
+  destroyDocumentHandler,
+  destroyDocumentRouteConfig,
+} from '$/routes/api/v3/projects/versions/documents/destroy'
+import {
   getHandler,
   getRouteConfig,
 } from '$/routes/api/v3/projects/versions/documents/get'
@@ -50,6 +54,10 @@ export const documentsRouter = createRouter()
   .openapi(
     route(getOrCreateRouteConfig, API_ROUTES.v3.projects.documents.getOrCreate),
     getOrCreateHandler,
+  )
+  .openapi(
+    route(destroyDocumentRouteConfig, API_ROUTES.v3.projects.documents.destroy),
+    destroyDocumentHandler,
   )
   .openapi(
     route(runRouteConfig, API_ROUTES.v3.projects.documents.run),