+++ packages/sdks/python/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.8.0] - 2026-02-23
+
+### Added
+
+- Added `prompts.delete(path, options?)` method to soft-delete a document/prompt from a draft project version.
+
 ## [5.7.0] - 2026-02-17
 
 ### Added